Add "Translate Reports" github action
