+++ site/gatsby-site/src/scripts/Translator.js
